fix: new getNearestPackageJson() helper function to remove duplicate code
